requirements.txt
Pillow==5.2.0
image==1.5.28